+++ app/charts/bar/bars-stacked-state-props.ts
@@ -137,23 +137,24 @@ export const useBarsStackedStateData = (
   const plottableData = usePlottableData(observations, {
     getX,
   });
-  const { sortedPlottableData, plottableDataWide } = useMemo(() => {
+  const plottableDataWide = useMemo(() => {
     const plottableDataByY = group(plottableData, getY);
-    const plottableDataWide = getWideData({
+
+    return getWideData({
       dataGrouped: plottableDataByY,
       key: y.componentId,
       getAxisValue: getX,
       getSegment,
     });
-
-    return {
-      sortedPlottableData: sortData(plottableData, {
-        plottableDataWide,
-      }),
-      plottableDataWide,
-    };
-  }, [plottableData, getX, y.componentId, getY, getSegment, sortData]);
-  const data = useChartData(sortedPlottableData, {
+  }, [plottableData, getX, y.componentId, getY, getSegment]);
+  const sortPlottableData = useCallback(
+    (data: Observation[]) => {
+      return sortData(data, { plottableDataWide });
+    },
+    [plottableDataWide, sortData]
+  );
+  const data = useChartData(plottableData, {
+    sortData: sortPlottableData,
     chartConfig,
     timeRangeDimensionId: yDimension.id,
     getAxisValueAsDate: getYAsDate,
@@ -163,7 +164,6 @@ export const useBarsStackedStateData = (
 
   return {
     ...data,
-    allData: sortedPlottableData,
     plottableDataWide,
   };
 };