+++ app/charts/column/columns-stacked-state-props.ts
@@ -137,23 +137,24 @@ export const useColumnsStackedStateData = (
   const plottableData = usePlottableData(observations, {
     getY,
   });
-  const { sortedPlottableData, plottableDataWide } = useMemo(() => {
+  const plottableDataWide = useMemo(() => {
     const plottableDataByX = group(plottableData, getX);
-    const plottableDataWide = getWideData({
+
+    return getWideData({
       dataGrouped: plottableDataByX,
       key: x.componentId,
       getAxisValue: getY,
       getSegment,
     });
-
-    return {
-      sortedPlottableData: sortData(plottableData, {
-        plottableDataWide,
-      }),
-      plottableDataWide,
-    };
-  }, [plottableData, getX, x.componentId, getY, getSegment, sortData]);
-  const data = useChartData(sortedPlottableData, {
+  }, [plottableData, getX, x.componentId, getY, getSegment]);
+  const sortPlottableData = useCallback(
+    (data: Observation[]) => {
+      return sortData(data, { plottableDataWide });
+    },
+    [plottableDataWide, sortData]
+  );
+  const data = useChartData(plottableData, {
+    sortData: sortPlottableData,
     chartConfig,
     timeRangeDimensionId: xDimension.id,
     getAxisValueAsDate: getXAsDate,
@@ -163,7 +164,6 @@ export const useColumnsStackedStateData = (
 
   return {
     ...data,
-    allData: sortedPlottableData,
     plottableDataWide,
   };
 };