+++ app/config-utils.ts
@@ -233,12 +233,12 @@ export const getMaybeValidChartConfigLimit = ({
   wouldBeValid: boolean;
 } => {
   const relatedAxisDimensionValue = axisDimension
-    ? limit.related.find((r) => r.dimensionId === axisDimension.id)
-        ?.dimensionValue
+    ? limit.related.find((r) => r.dimensionId === axisDimension.id)?.value
     : undefined;
   const relatedAxisDimensionValueLabel = axisDimension
-    ? axisDimension.values.find((v) => v.value === relatedAxisDimensionValue)
-        ?.label
+    ? [...axisDimension.values, ...axisDimension.relatedLimitValues].find(
+        (v) => v.value === relatedAxisDimensionValue
+      )?.label
     : undefined;
   const relatedToFilterBys = axisDimension
     ? limit.related.filter((r) => r.dimensionId !== axisDimension.id)
@@ -251,7 +251,7 @@ export const getMaybeValidChartConfigLimit = ({
         ? maybeFilter.value
         : undefined;
 
-    if (maybeFilterValue !== relatedToFilterBy.dimensionValue) {
+    if (maybeFilterValue !== relatedToFilterBy.value) {
       return {
         limit: undefined,
         relatedAxisDimensionValueLabel,
@@ -266,10 +266,7 @@ export const getMaybeValidChartConfigLimit = ({
     limit: measureConfigLimits.find((configLimit) => {
       return configLimit.related.every((cr) => {
         return limit.related.some((lr) => {
-          return (
-            lr.dimensionId === cr.dimensionId &&
-            lr.dimensionValue === cr.dimensionValue
-          );
+          return lr.dimensionId === cr.dimensionId && lr.value === cr.value;
         });
       });
     }),