+++ app/rdf/limits.ts
@@ -1,15 +1,18 @@
 import { CubeDimension } from "rdf-cube-view-query";
+import ParsingClient from "sparql-http-client/ParsingClient";
 
+import { DimensionValue } from "@/domain/data";
 import { truthy } from "@/domain/types";
 import { stringifyComponentId } from "@/graphql/make-component-id";
 import * as ns from "@/rdf/namespace";
+import { buildLocalizedSubQuery } from "@/rdf/query-utils";
 
 type BaseLimit = {
   name: string;
-  related: {
+  related: ({
     dimensionId: string;
-    dimensionValue: string;
-  }[];
+    value: string;
+  } & Pick<DimensionValue, "label" | "position">)[];
 };
 
 type LimitSingle = BaseLimit & {
@@ -25,21 +28,25 @@ type LimitRange = BaseLimit & {
 
 export type Limit = LimitSingle | LimitRange;
 
-export const getDimensionLimits = (
+export const getDimensionLimits = async (
   dim: CubeDimension,
-  { locale, unversionedCubeIri }: { locale: string; unversionedCubeIri: string }
-): Limit[] => {
-  return dim
+  {
+    locale,
+    unversionedCubeIri,
+    sparqlClient,
+  }: { locale: string; unversionedCubeIri: string; sparqlClient: ParsingClient }
+): Promise<Limit[]> => {
+  const baseLimits = dim
     .out(ns.cubeMeta.annotation)
-    .map((a) => {
+    .map((a, index) => {
       const name = a.out(ns.schema.name, { language: locale }).value ?? "";
       const ctxs = a.out(ns.cubeMeta.annotationContext);
       const related = ctxs
         .map((ctx) => {
           const dimensionIri = ctx.out(ns.sh.path).value;
-          const dimensionValue = ctx.out(ns.sh.hasValue).value;
+          const value = ctx.out(ns.sh.hasValue).value;
 
-          if (!dimensionIri || !dimensionValue) {
+          if (!dimensionIri || !value) {
             return null;
           }
 
@@ -48,7 +55,7 @@ export const getDimensionLimits = (
               unversionedCubeIri,
               unversionedComponentIri: dimensionIri,
             }),
-            dimensionValue,
+            value,
           };
         })
         .filter(truthy);
@@ -57,10 +64,13 @@ export const getDimensionLimits = (
 
       if (!!value) {
         return {
-          type: "single" as const,
-          name,
+          index,
           related,
-          value: +value,
+          limit: {
+            type: "single" as const,
+            name,
+            value: +value,
+          },
         };
       }
 
@@ -69,14 +79,70 @@ export const getDimensionLimits = (
 
       if (!!minValue && !!maxValue) {
         return {
-          type: "range" as const,
-          name,
+          index,
           related,
-          from: +minValue,
-          to: +maxValue,
+          limit: {
+            type: "range" as const,
+            name,
+            from: +minValue,
+            to: +maxValue,
+          },
         };
       }
     })
     .filter(truthy)
-    .sort((a, b) => a.name.localeCompare(b.name));
+    .sort((a, b) => a.limit.name.localeCompare(b.limit.name));
+
+  const baseRelated = baseLimits.flatMap((l) =>
+    l.related.map((r) => ({ ...r, index: l.index }))
+  );
+
+  if (baseRelated.length > 0) {
+    const allRelatedQuery = `PREFIX schema: <http://schema.org/>
+
+    SELECT ?index ?dimensionId ?value ?label ?position WHERE {
+      VALUES (?index ?dimensionId ?value) { ${baseLimits
+        .flatMap((l) => l.related.map((r) => ({ ...r, index: l.index })))
+        .map((r) => `(${r.index} <${r.dimensionId}> <${r.value}>)`)
+        .join(" ")} }
+        ${buildLocalizedSubQuery("value", "schema:name", "label", {
+          locale,
+        })}
+        ?value schema:position ?position .
+    }`;
+
+    const allRelated = (
+      await sparqlClient.query.select(allRelatedQuery, {
+        operation: "postUrlencoded",
+      })
+    ).map((d) => {
+      const index = +d.index.value;
+      const dimensionId = d.dimensionId.value;
+      const value = d.value.value;
+      const label = d.label.value;
+      const position = d.position.value;
+
+      return {
+        index,
+        dimensionId,
+        value,
+        label,
+        position,
+      };
+    });
+
+    return baseLimits.map(({ index, limit }) => {
+      return {
+        ...limit,
+        related: allRelated.filter((r) => r.index === index),
+      };
+    });
+  } else {
+    return baseLimits.map(({ limit }) => {
+      return {
+        ...limit,
+        related: [],
+      };
+    });
+  }
 };